swarm/chunk: address pr comments
